refactor: move DbTool type to db-common
