+++ bin/reth/src/commands/db/diff.rs
@@ -1,11 +1,11 @@
 use crate::{
     args::DatabaseArgs,
     dirs::{DataDirPath, PlatformPath},
-    utils::DbTool,
 };
 use clap::Parser;
 use reth_db::{open_db_read_only, tables_to_generic, DatabaseEnv, Tables};
 use reth_db_api::{cursor::DbCursorRO, database::Database, table::Table, transaction::DbTx};
+use reth_db_common::DbTool;
 use std::{
     collections::HashMap,
     fmt::Debug,
