+++ bin/reth/src/commands/db/list.rs
@@ -1,9 +1,9 @@
 use super::tui::DbListTUI;
-use crate::utils::{DbTool, ListFilter};
 use clap::Parser;
 use eyre::WrapErr;
 use reth_db::{DatabaseEnv, RawValue, TableViewer, Tables};
 use reth_db_api::{database::Database, table::Table};
+use reth_db_common::{DbTool, ListFilter};
 use reth_primitives::hex;
 use std::{cell::RefCell, sync::Arc};
 use tracing::error;
