+++ bin/reth/src/commands/db/stats.rs
@@ -1,4 +1,4 @@
-use crate::{commands::db::checksum::ChecksumViewer, utils::DbTool};
+use crate::commands::db::checksum::ChecksumViewer;
 use clap::Parser;
 use comfy_table::{Cell, Row, Table as ComfyTable};
 use eyre::WrapErr;
@@ -6,6 +6,7 @@ use human_bytes::human_bytes;
 use itertools::Itertools;
 use reth_db::{mdbx, static_file::iter_static_files, DatabaseEnv, TableViewer, Tables};
 use reth_db_api::database::Database;
+use reth_db_common::DbTool;
 use reth_fs_util as fs;
 use reth_node_core::dirs::{ChainPath, DataDirPath};
 use reth_provider::providers::StaticFileProvider;
