+++ bin/reth/src/commands/stage/dump/hashing_account.rs
@@ -1,8 +1,8 @@
 use super::setup;
-use crate::utils::DbTool;
 use eyre::Result;
 use reth_db::{tables, DatabaseEnv};
 use reth_db_api::{database::Database, table::TableImporter};
+use reth_db_common::DbTool;
 use reth_node_core::dirs::{ChainPath, DataDirPath};
 use reth_primitives::BlockNumber;
 use reth_provider::{providers::StaticFileProvider, ProviderFactory};
